Move to an appropriate directory
